--- a/README.md
+++ b/README.md
@@ -12,7 +12,6 @@ report bugs and feature requests on GitHub using the above URL.
 
 ### hub:
 - `csdiff`
-- `file`
 - `gzip`
 - `httpd`
 - `koji`
@@ -33,6 +32,7 @@ report bugs and feature requests on GitHub using the above URL.
 
 ### worker:
 - `csmock`
+- `file`
 - `koji`
 - `python3-kobo-client`
 - `python3-kobo-rpmlib`