--- a/containers/client.Dockerfile
+++ b/containers/client.Dockerfile
@@ -10,7 +10,6 @@ ENV OSH_CLIENT_CONFIG_FILE=osh/client/client-local.conf
 RUN echo -e "max_parallel_downloads=20\nfastestmirror=True" >> /etc/dnf/dnf.conf
 
 RUN dnf -y --setopt=tsflags=nodocs install \
-    file \
     koji \
     gzip \
     python3-coverage \