--- a/containers/client.Dockerfile
+++ b/containers/client.Dockerfile
@@ -9,11 +9,8 @@ ENV OSH_CLIENT_CONFIG_FILE=osh/client/client-local.conf
 
 RUN dnf -y --setopt=tsflags=nodocs install \
     koji \
-    gzip \
     python3-coverage \
-    python3-koji \
-    python36 \
-    xz
+    python36
 
 # store coverage to a separate volume
 RUN printf '[run]\ndata_file = /cov/coverage\n' > /coveragerc