--- a/containers/worker.Dockerfile
+++ b/containers/worker.Dockerfile
@@ -6,20 +6,24 @@ RUN dnf config-manager --set-enabled powertools
 WORKDIR /src
 
 ENV PYTHONPATH=.:kobo
-ENV OSH_WORKER_CONFIG_FILE=osh/worker/worker-local.conf
 
 RUN dnf -y --setopt=tsflags=nodocs install \
-    csmock \
-    file \
     koji \
     python3-coverage \
     python3
 
-RUN adduser csmock -G mock
-
 # store coverage to a separate volume
 RUN printf '[run]\ndata_file = /cov/coverage\n' > /coveragerc
 
+### END OF COMMON PART
+
+RUN dnf -y --setopt=tsflags=nodocs install \
+    csmock \
+    file
+
+RUN adduser csmock -G mock
+
+ENV OSH_WORKER_CONFIG_FILE=osh/worker/worker-local.conf
 RUN touch /WORKER_IS_READY
 
 CMD coverage-3 run --parallel-mode --omit="*site-packages*,*kobo*," --rcfile=/coveragerc osh/worker/osh-worker -f