Minor improvements that should make the container builds a bit faster:
```console 
# Before
$ ./containers/scripts/deploy.sh --clean
$ time ./containers/scripts/deploy.sh --full-dev
...
66.46s user 35.73s system 36% cpu 4:37.33 total

# After
$ ./containers/scripts/deploy.sh --clean
$ time ./containers/scripts/deploy.sh --full-dev
...
67.64s user 36.66s system 41% cpu 4:14.20 total
```